feat(http): split interface into ApiEventInterface and WebEventInterface; adjust routing and path prefixes - Introduce separate API and Web HTTP interfaces, with router prefixing and include_in_schema control; update plugin integration to use interface path_prefix and include_router options. - Refactor plugins to register with the appropriate interface: API endpoints (core, spacestate, smibhid) under “/api”, web endpoints (static files, docs) under root; docs now served at “/api/docs” and “/api/redoc”. - Wire the app to instantiate both interfaces, register them in the plugin lifecycle, and finalise setup independently. - Update configuration so the webserver path prefix applies to all endpoints (default “/”); adjust reverse proxy labels to match.
