+++ src/plugins/space/smibhid/__init__.py
@@ -2,12 +2,13 @@ __display_name__ = "S.M.I.B.H.I.D."
 __description__ = "A plugin to contain SMIBHID specific interfaces"
 __author__ = "Sam Cork"
 
-from smib.events.interfaces.http_event_interface import HttpEventInterface
+from smib.events.interfaces.http.http_api_event_interface import ApiEventInterface
 
 
-def register(http: HttpEventInterface):
+
+def register(api: ApiEventInterface):
     from .ui import register as register_ui_log_handlers
     from .sensor import register as register_sensor_log_handlers
 
-    register_ui_log_handlers(http)
-    register_sensor_log_handlers(http)
+    register_ui_log_handlers(api)
+    register_sensor_log_handlers(api)
