+++ src/smib/__main__.py
@@ -1,7 +1,6 @@
 import asyncio
 import logging
 from asyncio import CancelledError
-from pprint import pprint
 
 from pymongo.errors import PyMongoError
 from slack_bolt.app.async_app import AsyncApp
@@ -10,13 +9,13 @@ from smib.db.manager import DatabaseManager
 from smib.error_handler import slack_bolt_error_handler
 from smib.events.handlers.http_event_handler import HttpEventHandler
 from smib.events.handlers.scheduled_event_handler import ScheduledEventHandler
-from smib.events.interfaces.http_event_interface import HttpEventInterface
+from smib.events.interfaces.http.http_api_event_interface import ApiEventInterface
+from smib.events.interfaces.http.http_web_event_interface import WebEventInterface
 from smib.events.interfaces.scheduled_event_interface import ScheduledEventInterface
 from smib.events.services import EventServiceManager
 from smib.events.services.http_event_service import HttpEventService
 from smib.events.services.scheduled_event_service import ScheduledEventService
 from smib.events.services.slack_event_service import SlackEventService
-from smib.logging_ import initialise_logging
 from smib.plugins.integrations.database_plugin_integration import DatabasePluginIntegration
 from smib.plugins.integrations.http_plugin_integration import HttpPluginIntegration
 from smib.plugins.integrations.scheduled_plugin_integration import ScheduledPluginIntegration
@@ -55,7 +54,9 @@ async def main():
     # HTTP Service
     http_event_service = HttpEventService()
     http_event_handler = HttpEventHandler(bolt_app)
-    http_event_interface = HttpEventInterface(bolt_app, http_event_handler, http_event_service)
+
+    http_api_event_interface = ApiEventInterface(bolt_app, http_event_handler, http_event_service)
+    http_web_event_interface = WebEventInterface(bolt_app, http_event_handler, http_event_service)
 
     # Scheduled Job Service
     scheduled_event_service = ScheduledEventService()
@@ -75,27 +76,34 @@ async def main():
     plugin_lifecycle_manager = PluginLifecycleManager(bolt_app, plugin_loader)
 
     plugin_lifecycle_manager.register_parameter('slack', bolt_app)
-    plugin_lifecycle_manager.register_parameter('http', http_event_interface)
+    plugin_lifecycle_manager.register_parameter('http', http_api_event_interface)
+    plugin_lifecycle_manager.register_parameter('web', http_web_event_interface)
+    plugin_lifecycle_manager.register_parameter('api', http_api_event_interface)
     plugin_lifecycle_manager.register_parameter('schedule', scheduled_event_interface)
     plugin_lifecycle_manager.register_parameter('database', database_manager)
 
     # Plugin integrations
     slack_plugin_integration: SlackPluginIntegration = SlackPluginIntegration(bolt_app)
-    http_plugin_integration: HttpPluginIntegration = HttpPluginIntegration(http_event_interface)
+    http_api_plugin_integration: HttpPluginIntegration = HttpPluginIntegration(http_api_event_interface)
+    http_web_plugin_integration: HttpPluginIntegration = HttpPluginIntegration(http_web_event_interface)
     scheduled_plugin_integration: ScheduledPluginIntegration = ScheduledPluginIntegration(scheduled_event_interface)
     database_plugin_integration: DatabasePluginIntegration = DatabasePluginIntegration(plugin_lifecycle_manager)
 
     # Plugin lifecycle callbacks
     plugin_lifecycle_manager.register_plugin_unregister_callback(slack_plugin_integration.disconnect_plugin)
-    plugin_lifecycle_manager.register_plugin_unregister_callback(http_plugin_integration.disconnect_plugin)
+    plugin_lifecycle_manager.register_plugin_unregister_callback(http_api_plugin_integration.disconnect_plugin)
+    plugin_lifecycle_manager.register_plugin_unregister_callback(http_web_plugin_integration.disconnect_plugin)
     plugin_lifecycle_manager.register_plugin_unregister_callback(scheduled_plugin_integration.disconnect_plugin)
 
-    plugin_lifecycle_manager.register_plugin_preregister_callback(http_plugin_integration.initialise_plugin_router)
-    plugin_lifecycle_manager.register_plugin_postregister_callback(http_plugin_integration.remove_router_if_unused)
+    plugin_lifecycle_manager.register_plugin_preregister_callback(http_api_plugin_integration.initialise_plugin_router)
+    plugin_lifecycle_manager.register_plugin_postregister_callback(http_api_plugin_integration.remove_router_if_unused)
+    plugin_lifecycle_manager.register_plugin_preregister_callback(http_web_plugin_integration.initialise_plugin_router)
+    plugin_lifecycle_manager.register_plugin_postregister_callback(http_web_plugin_integration.remove_router_if_unused)
 
     plugin_lifecycle_manager.load_plugins()
 
-    http_plugin_integration.finalise_http_setup()
+    http_api_plugin_integration.finalise_http_setup()
+    http_web_plugin_integration.finalise_http_setup()
 
     database_manager.register_document_filter(database_plugin_integration.filter_valid_plugins)
