+++ src/smib/events/interfaces/http/http_web_event_interface.py
@@ -1,0 +1,14 @@
+from slack_bolt.app.async_app import AsyncApp
+
+from smib.events.handlers.http_event_handler import HttpEventHandler
+from smib.events.interfaces.http import HttpEventInterface
+from smib.events.services.http_event_service import HttpEventService
+
+
+class WebEventInterface(HttpEventInterface):
+    def __init__(self, bolt_app: AsyncApp, handler: HttpEventHandler, service: HttpEventService):
+        super().__init__(bolt_app, handler, service)
+
+    def get(self, path: str, *args, **kwargs):
+        """ See fastapi.FastAPI.get() for parameters """
+        return self._route_decorator(path, ["GET"], *args, **kwargs)
